Extract `Account::Mappings` concern from "interactions"
