+++ spec/models/concerns/account/interactions_spec.rb
@@ -4,94 +4,7 @@ require 'rails_helper'
 
 RSpec.describe Account::Interactions do
   let(:account)            { Fabricate(:account) }
-  let(:account_id)         { account.id }
-  let(:account_ids)        { [account_id] }
   let(:target_account)     { Fabricate(:account) }
-  let(:target_account_id)  { target_account.id }
-  let(:target_account_ids) { [target_account_id] }
-
-  describe '.following_map' do
-    subject { Account.following_map(target_account_ids, account_id) }
-
-    context 'when Account with Follow' do
-      it 'returns { target_account_id => true }' do
-        Fabricate(:follow, account: account, target_account: target_account)
-        expect(subject).to eq(target_account_id => { reblogs: true, notify: false, languages: nil })
-      end
-    end
-
-    context 'when Account without Follow' do
-      it 'returns {}' do
-        expect(subject).to eq({})
-      end
-    end
-  end
-
-  describe '.followed_by_map' do
-    subject { Account.followed_by_map(target_account_ids, account_id) }
-
-    context 'when Account with Follow' do
-      it 'returns { target_account_id => true }' do
-        Fabricate(:follow, account: target_account, target_account: account)
-        expect(subject).to eq(target_account_id => true)
-      end
-    end
-
-    context 'when Account without Follow' do
-      it 'returns {}' do
-        expect(subject).to eq({})
-      end
-    end
-  end
-
-  describe '.blocking_map' do
-    subject { Account.blocking_map(target_account_ids, account_id) }
-
-    context 'when Account with Block' do
-      it 'returns { target_account_id => true }' do
-        Fabricate(:block, account: account, target_account: target_account)
-        expect(subject).to eq(target_account_id => true)
-      end
-    end
-
-    context 'when Account without Block' do
-      it 'returns {}' do
-        expect(subject).to eq({})
-      end
-    end
-  end
-
-  describe '.muting_map' do
-    subject { Account.muting_map(target_account_ids, account_id) }
-
-    context 'when Account with Mute' do
-      before do
-        Fabricate(:mute, target_account: target_account, account: account, hide_notifications: hide)
-      end
-
-      context 'when Mute#hide_notifications?' do
-        let(:hide) { true }
-
-        it 'returns { target_account_id => { notifications: true } }' do
-          expect(subject).to eq(target_account_id => { notifications: true })
-        end
-      end
-
-      context 'when not Mute#hide_notifications?' do
-        let(:hide) { false }
-
-        it 'returns { target_account_id => { notifications: false } }' do
-          expect(subject).to eq(target_account_id => { notifications: false })
-        end
-      end
-    end
-
-    context 'when Account without Mute' do
-      it 'returns {}' do
-        expect(subject).to eq({})
-      end
-    end
-  end
 
   describe '#follow!' do
     it 'creates and returns Follow' do
